Adicionado senha para excluir todos registros

+++ src/main/java/com/viniciusx/proj_imc_on/service/ImcService.java
@@ -2,6 +2,8 @@ package com.viniciusx.proj_imc_on.service;
 
 import com.viniciusx.proj_imc_on.model.Imc;
 import com.viniciusx.proj_imc_on.repository.ImcRepository;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -41,8 +43,13 @@ public class ImcService {
         repository.deleteById(id);
     }
 
-    public void deleteAllImcs() {
-        repository.deleteAll();
-        repository.resetarSequencia();
+    public boolean deleteAllImcs(String senha) {
+        if (senha.equals("A1b2c3@")) {
+            repository.deleteAll();
+            return true;
+        } else {
+            return false;
+        }
+
     }
 }
